backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
asyncpg
PyJWT
email-validator
bcrypt
python-jose
jose